+++ Assets/Scripts/Ammunition/Guns/RocketLauncher.cs
@@ -9,7 +9,7 @@ namespace MyGame
 	{
 		public Rocket m_rocket;
 		public float speed { get; set; }
-		public float demage { get; set; }
+		public int demage { get; set; }
 		public float factor { get; set; }
 
 		protected override void DoAfterInit()
